# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Suno Content",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/atas/{identificador}": {
      "get": {
        "summary": "Obter Ata",
        "operationId": "obter_ata_api_atas__identificador__get",
        "parameters": [
          {
            "name": "identificador",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identificador"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ata"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/execucoes/{identificador}": {
      "get": {
        "summary": "Obter Execucao",
        "operationId": "obter_execucao_api_execucoes__identificador__get",
        "parameters": [
          {
            "name": "identificador",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identificador"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Execucao"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Renomear Execucao",
        "operationId": "renomear_execucao_api_execucoes__identificador__patch",
        "parameters": [
          {
            "name": "identificador",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identificador"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Renomeacao"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecucaoResumo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/execucoes/{identificador}/resumo": {
      "get": {
        "summary": "Obter Resumo",
        "operationId": "obter_resumo_api_execucoes__identificador__resumo_get",
        "parameters": [
          {
            "name": "identificador",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identificador"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecucaoResumo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/execucoes/{identificador}/celulas/{audiencia}/{formato}/decisao": {
      "post": {
        "summary": "Registrar Decisao",
        "operationId": "registrar_decisao_api_execucoes__identificador__celulas__audiencia___formato__decisao_post",
        "parameters": [
          {
            "name": "identificador",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identificador"
            }
          },
          {
            "name": "audiencia",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Audiencia"
            }
          },
          {
            "name": "formato",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Formato"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PedidoDeDecisao"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisaoHumana"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/execucoes/{identificador}/exportar": {
      "get": {
        "summary": "Exportar Execucao",
        "operationId": "exportar_execucao_api_execucoes__identificador__exportar_get",
        "parameters": [
          {
            "name": "identificador",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identificador"
            }
          },
          {
            "name": "formato",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/FormatoExportacao"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/execucoes/{identificador}/ancoras": {
      "get": {
        "summary": "Obter Ancoras",
        "operationId": "obter_ancoras_api_execucoes__identificador__ancoras_get",
        "parameters": [
          {
            "name": "identificador",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identificador"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ancoras"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/execucoes/{identificador}/celulas/{audiencia}/{formato}": {
      "get": {
        "summary": "Obter Celula",
        "operationId": "obter_celula_api_execucoes__identificador__celulas__audiencia___formato__get",
        "parameters": [
          {
            "name": "identificador",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identificador"
            }
          },
          {
            "name": "audiencia",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Audiencia"
            }
          },
          {
            "name": "formato",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Formato"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoricoCelula"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/execucoes/{identificador}/celulas/{audiencia}/{formato}/ancoras": {
      "get": {
        "summary": "Obter Ancoras Da Celula",
        "description": "As Âncoras numéricas citadas pela Célula final — \"Âncoras ao lado da Célula\".",
        "operationId": "obter_ancoras_da_celula_api_execucoes__identificador__celulas__audiencia___formato__ancoras_get",
        "parameters": [
          {
            "name": "identificador",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identificador"
            }
          },
          {
            "name": "audiencia",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Audiencia"
            }
          },
          {
            "name": "formato",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Formato"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AncoraNumerica"
                  },
                  "title": "Response Obter Ancoras Da Celula Api Execucoes  Identificador  Celulas  Audiencia   Formato  Ancoras Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/execucoes/{identificador}/pacotes": {
      "get": {
        "summary": "Listar Pacotes",
        "operationId": "listar_pacotes_api_execucoes__identificador__pacotes_get",
        "parameters": [
          {
            "name": "identificador",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identificador"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PacotePublicacao"
                  },
                  "title": "Response Listar Pacotes Api Execucoes  Identificador  Pacotes Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/execucoes/{identificador}/arquivos/{caminho}": {
      "get": {
        "summary": "Obter Arquivo",
        "operationId": "obter_arquivo_api_execucoes__identificador__arquivos__caminho__get",
        "parameters": [
          {
            "name": "identificador",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identificador"
            }
          },
          {
            "name": "caminho",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Caminho"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Ata": {
        "properties": {
          "identificador": {
            "type": "string",
            "title": "Identificador",
            "description": "Nome estável do arquivo em data/atas/, ex. 'copom-280-2026-08-05'."
          },
          "titulo": {
            "type": "string",
            "title": "Titulo"
          },
          "reuniao": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reuniao",
            "description": "Número da reunião do Copom."
          },
          "data_referencia": {
            "type": "string",
            "format": "date",
            "title": "Data Referencia"
          },
          "origem_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origem Url",
            "description": "URL de onde o arquivo foi buscado, quando foi."
          },
          "arquivo": {
            "anyOf": [
              {
                "type": "string",
                "format": "path"
              },
              {
                "type": "null"
              }
            ],
            "title": "Arquivo",
            "description": "Caminho local do arquivo original."
          },
          "texto": {
            "type": "string",
            "title": "Texto",
            "description": "Texto já extraído; é o que o Gerador lê."
          },
          "idioma": {
            "type": "string",
            "enum": [
              "pt",
              "en"
            ],
            "title": "Idioma",
            "default": "pt"
          }
        },
        "type": "object",
        "required": [
          "identificador",
          "titulo",
          "data_referencia",
          "texto"
        ],
        "title": "Ata",
        "description": "O comunicado financeiro público que entra no sistema (ADR 0006)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  }
}